requirements.txt
# LangChain核心库
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-classic>=0.0.1

# 向量数据库
faiss-cpu>=1.7.4

# Embedding模型
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0

# LLM提供商
zhipuai>=2.0.0

# 文档处理
python-docx>=1.0.0

# 文本分割
tiktoken>=0.5.0

# 环境变量管理
python-dotenv>=1.0.0

# 基础依赖
pydantic>=2.0.0
